9P: send RFLUSH only after reply to a request.
The previous FLUSH mechanism was: do not send the reply to a request if a TFLUSH was received.

This does not work for TWALK, because it will create a new fid that does not exist on the client, since the client considers that the TWALK has been aborted.

Now we always send the reply to a request, and send the RFLUSH (if any) afterwards.